test: split isIdentityCard validator tests
