+++ test/validators.test.js
@@ -114,423 +114,6 @@ describe('Validators', () => {
     });
   });
 
-  it('should validate identity cards', () => {
-    const fixtures = [
-      {
-        locale: 'PK',
-        valid: [
-          '45504-4185771-3',
-          '39915-6182971-9',
-          '21143-6182971-2',
-          '34543-2323471-1',
-          '72345-2345678-7',
-          '63456-8765432-8',
-          '55672-1234567-5',
-          '21234-9876543-6',
-        ],
-        invalid: [
-          '08000-1234567-5',
-          '74321-87654321-1',
-          '51234-98765-2',
-          '00000-0000000-0',
-          '88888-88888888-0',
-          '99999-9999999-9',
-          '11111',
-        ],
-      },
-      {
-        locale: 'zh-HK',
-        valid: [
-          'OV290326[A]',
-          'Q803337[0]',
-          'Z0977986',
-          'W520128(7)',
-          'A494866[4]',
-          'A494866(4)',
-          'Z867821A',
-          'ag293013(9)',
-          'k348609(5)',
-        ],
-        invalid: [
-          'A1234567890',
-          '98765432',
-          'O962472(9)',
-          'M4578601',
-          'X731324[8]',
-          'C503134(5)',
-          'RH265886(3)',
-        ],
-      },
-      {
-        locale: 'LK',
-        valid: [
-          '722222222v',
-          '722222222V',
-          '993151225x',
-          '993151225X',
-          '188888388x',
-          '935632124V',
-          '199931512253',
-          '200023125632',
-        ],
-        invalid: [
-          '023125648V',
-          '023345621v',
-          '021354211X',
-          '055321231x',
-          '02135465462',
-          '199931512253X',
-        ],
-      },
-      {
-        locale: 'PL',
-        valid: [
-          '99012229019',
-          '09210215408',
-          '20313034701',
-          '86051575214',
-          '77334586883',
-          '54007481320',
-          '06566860643',
-          '77552478861',
-        ],
-        invalid: [
-          'aa',
-          '5',
-          '195',
-          '',
-          ' ',
-          '12345678901',
-          '99212229019',
-          '09210215402',
-          '20313534701',
-          '86241579214',
-        ],
-      },
-      {
-        locale: 'ES',
-        valid: [
-          '99999999R',
-          '12345678Z',
-          '01234567L',
-          '01234567l',
-          'X1234567l',
-          'x1234567l',
-          'X1234567L',
-          'Y1234567X',
-          'Z1234567R',
-        ],
-        invalid: [
-          '123456789',
-          '12345678A',
-          '12345 678Z',
-          '12345678-Z',
-          '1234*6789',
-          '1234*678Z',
-          '12345678!',
-          '1234567L',
-          'A1234567L',
-          'X1234567A',
-          'Y1234567B',
-          'Z1234567C',
-        ],
-      },
-      {
-        locale: 'FI',
-        valid: [
-          '131052-308T', // People born in 1900s
-          '131052A308T', // People born in 2000s
-          '131052+308T', // People born in 1800s
-          '131052-313Y',
-        ],
-        invalid: [
-          '131052308T',
-          '131052-308T ',
-          '131052-308A',
-        ],
-      },
-      {
-        locale: 'IN',
-        valid: [
-          '298448863364',
-          '2984 4886 3364',
-        ],
-        invalid: [
-          '99999999R',
-          '12345678Z',
-          '01234567L',
-          '01234567l',
-          'X1234567l',
-          'x1234567l',
-          'X1234567L',
-        ],
-      },
-      {
-        locale: 'IR',
-        valid: [
-          '0499370899',
-          '0790419904',
-          '0084575948',
-          '0963695398',
-          '0684159414',
-          '0067749828',
-          '0650451252',
-          '1583250689',
-          '4032152314',
-          '0076229645',
-          '4271467685',
-          '0200203241',
-        ],
-        invalid: [
-          '1260293040',
-          '0000000001',
-          '1999999999',
-          '9999999991',
-          'AAAAAAAAAA',
-          '0684159415',
-        ],
-      },
-      {
-        locale: 'IT',
-        valid: [
-          'CR43675TM',
-          'CA79382RA',
-        ],
-        invalid: [
-          'CA00000AA',
-          'CB2342TG',
-          'CS123456A',
-          'C1236EC',
-        ],
-      },
-      {
-        locale: 'NO',
-        valid: [
-          '09053426694',
-          '26028338723',
-          '08031470790',
-          '12051539514',
-          '02077448074',
-          '14035638319',
-          '13031379673',
-          '29126214926',
-        ],
-        invalid: [
-          '09053426699',
-          '00000000000',
-          '26028338724',
-          '92031470790',
-        ],
-      },
-      {
-        locale: 'TH',
-        valid: [
-          '1101230000001',
-          '1101230000060',
-        ],
-        invalid: [
-          'abc',
-          '1101230',
-          '11012300000011',
-          'aaaaaaaaaaaaa',
-          '110123abcd001',
-          '1101230000007',
-          '0101123450000',
-          '0101123450004',
-          '9101123450008',
-        ],
-      },
-      {
-        locale: 'he-IL',
-        valid: [
-          '219472156',
-          '219486610',
-          '219488962',
-          '219566726',
-          '219640216',
-          '219645041',
-          '334795465',
-          '335211686',
-          '335240479',
-          '335472171',
-          '336999842',
-          '337090443',
-        ],
-        invalid: [
-          '123456789',
-          '12345678A',
-          '12345 678Z',
-          '12345678-Z',
-          '1234*6789',
-          '1234*678Z',
-          '12345678!',
-          '1234567L',
-          'A1234567L',
-          'X1234567A',
-          'Y1234567B',
-          'Z1234567C',
-          '219772156',
-          '219487710',
-          '334705465',
-          '336000842',
-        ],
-      },
-      {
-        locale: 'ar-LY',
-        valid: [
-          '119803455876',
-          '120024679875',
-          '219624876201',
-          '220103480657',
-        ],
-        invalid: [
-          '987654320123',
-          '123-456-7890',
-          '012345678912',
-          '1234567890',
-          'AFJBHUYTREWR',
-          'C4V6B1X0M5T6',
-          '9876543210123',
-        ],
-      },
-      {
-        locale: 'ar-TN',
-        valid: [
-          '09958092',
-          '09151092',
-          '65126506',
-          '79378815',
-          '58994407',
-          '73089789',
-          '73260311',
-        ],
-        invalid: [
-          '123456789546',
-          '123456789',
-          '023456789',
-          '12345678A',
-          '12345',
-          '1234578A',
-          '123 578A',
-          '12345 678Z',
-          '12345678-Z',
-          '1234*6789',
-          '1234*678Z',
-          'GE9800as98',
-          'X231071922',
-          '1234*678Z',
-          '12345678!',
-        ],
-      },
-      {
-        locale: 'zh-CN',
-        valid: [
-          '235407195106112745',
-          '210203197503102721',
-          '520323197806058856',
-          '110101491001001',
-        ],
-        invalid: [
-          '160323197806058856',
-          '010203197503102721',
-          '520323297806058856',
-          '520323197802318856',
-          '235407195106112742',
-          '010101491001001',
-          '110101491041001',
-          '160101491001001',
-          '110101940231001',
-          'xx1234567',
-          '135407195106112742',
-          '123456789546',
-          '123456789',
-          '023456789',
-          '12345678A',
-          '12345',
-          '1234578A',
-          '123 578A',
-          '12345 678Z',
-          '12345678-Z',
-          '1234*6789',
-          '1234*678Z',
-          'GE9800as98',
-          'X231071922',
-          '1234*678Z',
-          '12345678!',
-          '235407207006112742',
-        ],
-      },
-      {
-        locale: 'zh-TW',
-        valid: [
-          'B176944193',
-          'K101189797',
-          'F112866121',
-          'A219758834',
-          'A244144802',
-          'A146047171',
-          'Q170219004',
-          'Z277018381',
-          'X231071923',
-        ],
-        invalid: [
-          '123456789',
-          'A185034995',
-          'X431071923',
-          'GE9800as98',
-          'X231071922',
-          '1234*678Z',
-          '12345678!',
-          '1234567L',
-          'A1234567L',
-          'X1234567A',
-          'Y1234567B',
-          'Z1234567C',
-          '219772156',
-          '219487710',
-          '334705465',
-          '336000842',
-        ],
-      },
-    ];
-
-    let allValid = [];
-
-    // Test fixtures
-    fixtures.forEach((fixture) => {
-      if (fixture.valid) allValid = allValid.concat(fixture.valid);
-      test({
-        validator: 'isIdentityCard',
-        valid: fixture.valid,
-        invalid: fixture.invalid,
-        args: [fixture.locale],
-      });
-    });
-
-    // Test generics
-    test({
-      validator: 'isIdentityCard',
-      valid: [
-        ...allValid,
-      ],
-      invalid: [
-        'foo',
-      ],
-      args: ['any'],
-    });
-  });
-
-  it('should error on invalid locale', () => {
-    test({
-      validator: 'isIdentityCard',
-      args: ['is-NOT'],
-      error: [
-        '99999999R',
-        '12345678Z',
-      ],
-    });
-  });
-
   it('should validate ISINs', () => {
     test({
       validator: 'isISIN',
